services/kafka-consumers/rabbitmq-publisher/requirements.txt
pika>=1.3.2
kafka-python>=2.0.2
